+++ src/Config/ConfigManager.cpp
@@ -235,6 +235,8 @@ namespace gamescope::config
                 z.height = JGetFloat( *pZoom, "height", z.height );
                 z.factor = JGetFloat( *pZoom, "factor", z.factor );
                 z.mouse_scale = JGetBool( *pZoom, "mouse_scale", z.mouse_scale );
+                z.consume_button = JGetBool( *pZoom, "consume_button", z.consume_button );
+                z.scroll_adjust = JGetBool( *pZoom, "scroll_adjust", z.scroll_adjust );
             }
 
             if ( const nlohmann::json *pReshade = JGetObject( j, "reshade" ) )
@@ -562,6 +564,8 @@ namespace gamescope::config
             jZoom[ "height" ] = z.height;
             jZoom[ "factor" ] = z.factor;
             jZoom[ "mouse_scale" ] = z.mouse_scale;
+            jZoom[ "consume_button" ] = z.consume_button;
+            jZoom[ "scroll_adjust" ] = z.scroll_adjust;
 
             nlohmann::json jSaturation = nlohmann::json::object();
             jSaturation[ "enabled" ] = s.reshade.saturation.enabled;